requirements.txt
pillow
numpy
torch
torchvision
tqdm
opencv-python   
pytesseract 
matplotlib 
cascadetabnet
python-docx
pandas
tensorflow